governance_weekly/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.16.0
sqlalchemy>=2.0.25
readability-lxml>=0.8.1
reportlab>=4.0.9
google-cloud-translate>=3.12.1
transformers>=4.36.2
torch>=2.1.2  # For MarianMT/Transformers
python-dotenv>=1.0.0
gspread>=6.0.0 # For potential google sheet integration
google-auth>=2.26.1
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.111.0
schedule>=1.2.1
lxml>=5.1.0
fake-useragent>=1.4.0
